Delete InProcessPanelMatchResourceSetup
