+++ src/main/kotlin/org/wfanet/measurement/integration/common/InProcessKingdom.kt
@@ -63,7 +63,7 @@ import org.wfanet.measurement.kingdom.service.system.v1alpha.ComputationLogEntri
 import org.wfanet.measurement.kingdom.service.system.v1alpha.ComputationParticipantsService as SystemComputationParticipantsService
 import org.wfanet.measurement.kingdom.service.system.v1alpha.ComputationsService as SystemComputationsService
 import org.wfanet.measurement.kingdom.service.system.v1alpha.RequisitionsService as SystemRequisitionsService
-import org.wfanet.measurement.loadtest.panelmatchresourcesetup.InProcessPanelMatchResourceSetup
+import org.wfanet.measurement.loadtest.panelmatchresourcesetup.PanelMatchResourceSetup
 
 /** TestRule that starts and stops all Kingdom gRPC services. */
 class InProcessKingdom(
@@ -191,9 +191,9 @@ class InProcessKingdom(
   val systemApiChannel: Channel
     get() = systemApiServer.channel
 
-  /** Provides a InProcessPanelMatchResourceSetup instance with the Kingdom's internal API. */
-  val panelMatchResourceSetup: InProcessPanelMatchResourceSetup
-    get() = InProcessPanelMatchResourceSetup(internalApiChannel)
+  /** Provides a PanelMatchResourceSetup instance with the Kingdom's internal API. */
+  val panelMatchResourceSetup: PanelMatchResourceSetup
+    get() = PanelMatchResourceSetup(internalApiChannel)
 
   /** Provides access to Account and DataProvider creation in place of the Kingdom's operator. */
   val internalAccountsClient by lazy { InternalAccountsCoroutineStub(internalApiChannel) }
